update for special case of Eps Eri
this update allows for the optional use of hard-coded coefficients
for the special case of Epsilon Eridani. The reason for this is
because previous code that looked up the values in a table
failed when running the code using MPI. This was because multiple
processes were trying to restore the table simultaneously, making
rows go missing.